+++ dnsimple/registrar.go
@@ -19,7 +19,7 @@ type DomainCheck struct {
 	Premium   bool   `json:"premium"`
 }
 
-// DomainCheckResponse represents a response from the domain check.
+// DomainCheckResponse represents a response from a domain check request.
 type DomainCheckResponse struct {
 	Response
 	Data *DomainCheck `json:"data"`
@@ -28,7 +28,7 @@ type DomainCheckResponse struct {
 // CheckDomain checks a domain name.
 //
 // See https://developer.dnsimple.com/v2/registrar/#check
-func (s *RegistrarService) CheckDomain(accountID string, domainName string) (*DomainCheckResponse, error) {
+func (s *RegistrarService) CheckDomain(accountID, domainName string) (*DomainCheckResponse, error) {
 	path := versioned(fmt.Sprintf("/%v/registrar/domains/%v/check", accountID, domainName))
 	checkResponse := &DomainCheckResponse{}
 
@@ -41,6 +41,56 @@ func (s *RegistrarService) CheckDomain(accountID string, domainName string) (*Do
 	return checkResponse, nil
 }
 
+// DomainPremiumPrice represents the premium price for a premium domain.
+type DomainPremiumPrice struct {
+	// The domain premium price
+	PremiumPrice string `json:"premium_price"`
+	// The registrar action.
+	// Possible values are registration|transfer|renewal
+	Action string `json:"action"`
+}
+
+// DomainPremiumPriceResponse represents a response from a domain premium price request.
+type DomainPremiumPriceResponse struct {
+	Response
+	Data *DomainPremiumPrice `json:"data"`
+}
+
+// DomainPremiumPriceOptions specifies the optional parameters you can provide
+// to customize the RegistrarService.GetDomainPremiumPrice method.
+type DomainPremiumPriceOptions struct {
+	Action string `url:"action,omitempty"`
+}
+
+// Gets the premium price for a domain.
+//
+// You must specify an action to get the price for. Valid actions are:
+// - registration
+// - transfer
+// - renewal
+//
+// See https://developer.dnsimple.com/v2/registrar/#premium-price
+func (s *RegistrarService) GetDomainPremiumPrice(accountID, domainName string, options *DomainPremiumPriceOptions) (*DomainPremiumPriceResponse, error) {
+	var err error
+	path := versioned(fmt.Sprintf("/%v/registrar/domains/%v/premium_price", accountID, domainName))
+	priceResponse := &DomainPremiumPriceResponse{}
+
+	if options != nil {
+		path, err = addURLQueryOptions(path, options)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	resp, err := s.client.get(path, priceResponse)
+	if err != nil {
+		return nil, err
+	}
+
+	priceResponse.HttpResponse = resp
+	return priceResponse, nil
+}
+
 // DomainRegisterRequest represents the attributes you can pass to a register API request.
 // Some attributes are mandatory.
 type DomainRegisterRequest struct {
@@ -48,16 +98,30 @@ type DomainRegisterRequest struct {
 	RegistrantID int `json:"registrant_id"`
 	// Set to true to enable the whois privacy service. An extra cost may apply.
 	// Default to false.
-	EnableWhoisPrivacy bool `json:"private_whois,omitempty"`
+	EnableWhoisPrivacy bool `json:"whois_privacy,omitempty"`
 	// Set to true to enable the auto-renewal of the domain.
 	// Default to true.
 	EnableAutoRenewal bool `json:"auto_renew,omitempty"`
 }
 
+// DomainRegistration represents the result of a domain renewal call.
+type DomainRegistration struct {
+	ID           int    `json:"id"`
+	DomainID     int    `json:"domain_id"`
+	RegistrantID int    `json:"registrant_id"`
+	Period       int    `json:"period"`
+	State        string `json:"state"`
+	AutoRenew    bool   `json:"auto_renew"`
+	WhoisPrivacy bool   `json:"whois_privacy"`
+	PremiumPrice string `json:"premium_price"`
+	CreatedAt    string `json:"created_at,omitempty"`
+	UpdatedAt    string `json:"updated_at,omitempty"`
+}
+
 // DomainRegistrationResponse represents a response from an API method that results in a domain registration.
 type DomainRegistrationResponse struct {
 	Response
-	Data *Domain `json:"data"`
+	Data *DomainRegistration `json:"data"`
 }
 
 // RegisterDomain registers a domain name.
@@ -85,19 +149,32 @@ type DomainTransferRequest struct {
 	RegistrantID int `json:"registrant_id"`
 	// The Auth-Code required to transfer the domain.
 	// This is provided by the current registrar of the domain.
-	AuthInfo string `json:"auth_info,omitempty"`
+	AuthCode string `json:"auth_code,omitempty"`
 	// Set to true to enable the whois privacy service. An extra cost may apply.
 	// Default to false.
-	EnableWhoisPrivacy bool `json:"private_whois,omitempty"`
+	EnableWhoisPrivacy bool `json:"whois_privacy,omitempty"`
 	// Set to true to enable the auto-renewal of the domain.
 	// Default to true.
 	EnableAutoRenewal bool `json:"auto_renew,omitempty"`
 }
 
+// DomainTransfer represents the result of a domain renewal call.
+type DomainTransfer struct {
+	ID           int    `json:"id"`
+	DomainID     int    `json:"domain_id"`
+	RegistrantID int    `json:"registrant_id"`
+	State        string `json:"state"`
+	AutoRenew    bool   `json:"auto_renew"`
+	WhoisPrivacy bool   `json:"whois_privacy"`
+	PremiumPrice string `json:"premium_price"`
+	CreatedAt    string `json:"created_at,omitempty"`
+	UpdatedAt    string `json:"updated_at,omitempty"`
+}
+
 // DomainTransferResponse represents a response from an API method that results in a domain transfer.
 type DomainTransferResponse struct {
 	Response
-	Data *Domain `json:"data"`
+	Data *DomainTransfer `json:"data"`
 }
 
 // TransferDomain transfers a domain name.
@@ -147,10 +224,21 @@ type DomainRenewRequest struct {
 	Period int `json:"period"`
 }
 
-// DomainRenewalResponse represents a response from an API method that results in a domain renewal.
+// DomainRenewal represents the result of a domain renewal call.
+type DomainRenewal struct {
+	ID           int    `json:"id"`
+	DomainID     int    `json:"domain_id"`
+	Period       int    `json:"period"`
+	State        string `json:"state"`
+	PremiumPrice string `json:"premium_price"`
+	CreatedAt    string `json:"created_at,omitempty"`
+	UpdatedAt    string `json:"updated_at,omitempty"`
+}
+
+// DomainRenewalResponse represents a response from an API method that returns a domain renewal.
 type DomainRenewalResponse struct {
 	Response
-	Data *Domain `json:"data"`
+	Data *DomainRenewal `json:"data"`
 }
 
 // RenewDomain renews a domain name.