+++ dnsimple/registrar_test.go
@@ -3,6 +3,7 @@ package dnsimple
 import (
 	"io"
 	"net/http"
+	"net/url"
 	"testing"
 )
 
@@ -31,6 +32,55 @@ func TestRegistrarService_CheckDomain(t *testing.T) {
 	}
 }
 
+func TestRegistrarService_GetDomainPremiumPrice(t *testing.T) {
+	setupMockServer()
+	defer teardownMockServer()
+
+	mux.HandleFunc("/v2/1010/registrar/domains/example.com/premium_price", func(w http.ResponseWriter, r *http.Request) {
+		httpResponse := httpResponseFixture(t, "/getDomainPremiumPrice/success.http")
+
+		testMethod(t, r, "GET")
+		testHeaders(t, r)
+
+		w.WriteHeader(httpResponse.StatusCode)
+		io.Copy(w, httpResponse.Body)
+	})
+
+	priceResponse, err := client.Registrar.GetDomainPremiumPrice("1010", "example.com", nil)
+	if err != nil {
+		t.Fatalf("Registrar.GetDomainPremiumPrice() returned error: %v", err)
+	}
+
+	price := priceResponse.Data
+	if want, got := "109.00", price.PremiumPrice; want != got {
+		t.Fatalf("Registrar.GetDomainPremiumPrice() returned Domain expected to be `%v`, got `%v`", want, got)
+	}
+}
+
+func TestRegistrarService_GetDomainPremiumPrice_WithOptions(t *testing.T) {
+	setupMockServer()
+	defer teardownMockServer()
+
+	mux.HandleFunc("/v2/1010/registrar/domains/example.com/premium_price", func(w http.ResponseWriter, r *http.Request) {
+		httpResponse := httpResponseFixture(t, "/getDomainPremiumPrice/success.http")
+
+		testQuery(t, r, url.Values{
+			"action": []string{"registration"},
+		})
+
+		testMethod(t, r, "GET")
+		testHeaders(t, r)
+
+		w.WriteHeader(httpResponse.StatusCode)
+		io.Copy(w, httpResponse.Body)
+	})
+
+	_, err := client.Registrar.GetDomainPremiumPrice("1010", "example.com", &DomainPremiumPriceOptions{Action: "registration"})
+	if err != nil {
+		t.Fatalf("Registrar.GetDomainPremiumPrice() returned error: %v", err)
+	}
+}
+
 func TestRegistrarService_RegisterDomain(t *testing.T) {
 	setupMockServer()
 	defer teardownMockServer()
@@ -52,15 +102,15 @@ func TestRegistrarService_RegisterDomain(t *testing.T) {
 
 	registrationResponse, err := client.Registrar.RegisterDomain("1010", "example.com", registerRequest)
 	if err != nil {
-		t.Fatalf("Registrar.Register() returned error: %v", err)
+		t.Fatalf("Registrar.RegisterDomain() returned error: %v", err)
 	}
 
-	domain := registrationResponse.Data
-	if want, got := 1, domain.ID; want != got {
-		t.Fatalf("Registrar.Register() returned ID expected to be `%v`, got `%v`", want, got)
+	registration := registrationResponse.Data
+	if want, got := 1, registration.ID; want != got {
+		t.Fatalf("Registrar.RegisterDomain() returned ID expected to be `%v`, got `%v`", want, got)
 	}
-	if want, got := "example.com", domain.Name; want != got {
-		t.Fatalf("Registrar.Register() returned Name expected to be `%v`, got `%v`", want, got)
+	if want, got := 999, registration.DomainID; want != got {
+		t.Fatalf("Registrar.RegisterDomain() returned Name expected to be `%v`, got `%v`", want, got)
 	}
 }
 
@@ -74,26 +124,26 @@ func TestRegistrarService_TransferDomain(t *testing.T) {
 		testMethod(t, r, "POST")
 		testHeaders(t, r)
 
-		want := map[string]interface{}{"registrant_id": float64(2), "auth_info": "x1y2z3"}
+		want := map[string]interface{}{"registrant_id": float64(2), "auth_code": "x1y2z3"}
 		testRequestJSON(t, r, want)
 
 		w.WriteHeader(httpResponse.StatusCode)
 		io.Copy(w, httpResponse.Body)
 	})
 
-	transferRequest := &DomainTransferRequest{RegistrantID: 2, AuthInfo: "x1y2z3"}
+	transferRequest := &DomainTransferRequest{RegistrantID: 2, AuthCode: "x1y2z3"}
 
 	transferResponse, err := client.Registrar.TransferDomain("1010", "example.com", transferRequest)
 	if err != nil {
-		t.Fatalf("Registrar.Transfer() returned error: %v", err)
+		t.Fatalf("Registrar.TransferDomain() returned error: %v", err)
 	}
 
-	domain := transferResponse.Data
-	if want, got := 1, domain.ID; want != got {
-		t.Fatalf("Registrar.Transfer() returned ID expected to be `%v`, got `%v`", want, got)
+	transfer := transferResponse.Data
+	if want, got := 1, transfer.ID; want != got {
+		t.Fatalf("Registrar.TransferDomain() returned ID expected to be `%v`, got `%v`", want, got)
 	}
-	if want, got := "example.com", domain.Name; want != got {
-		t.Fatalf("Registrar.Transfer() returned Name expected to be `%v`, got `%v`", want, got)
+	if want, got := 999, transfer.DomainID; want != got {
+		t.Fatalf("Registrar.TransferDomain() returned Name expected to be `%v`, got `%v`", want, got)
 	}
 }
 
@@ -134,16 +184,16 @@ func TestRegistrarService_RenewDomain(t *testing.T) {
 		io.Copy(w, httpResponse.Body)
 	})
 
-	registrationResponse, err := client.Registrar.RenewDomain("1010", "example.com", nil)
+	renewalResponse, err := client.Registrar.RenewDomain("1010", "example.com", nil)
 	if err != nil {
-		t.Fatalf("Registrar.Renew() returned error: %v", err)
+		t.Fatalf("Registrar.RenewDomain() returned error: %v", err)
 	}
 
-	domain := registrationResponse.Data
-	if want, got := 1, domain.ID; want != got {
-		t.Fatalf("Registrar.Renew() returned ID expected to be `%v`, got `%v`", want, got)
+	renewal := renewalResponse.Data
+	if want, got := 1, renewal.ID; want != got {
+		t.Fatalf("Registrar.RenewDomain() returned ID expected to be `%v`, got `%v`", want, got)
 	}
-	if want, got := "example.com", domain.Name; want != got {
-		t.Fatalf("Registrar.Renew() returned Name expected to be `%v`, got `%v`", want, got)
+	if want, got := 999, renewal.DomainID; want != got {
+		t.Fatalf("Registrar.RenewDomain() returned DomainID expected to be `%v`, got `%v`", want, got)
 	}
 }