+++ dnsimple/templates.go
@@ -15,9 +15,9 @@ type TemplatesService struct {
 // Template represents a Template in DNSimple.
 type Template struct {
 	ID          int    `json:"id,omitempty"`
+	SID         string `json:"sid,omitempty"`
 	AccountID   int    `json:"account_id,omitempty"`
 	Name        string `json:"name,omitempty"`
-	ShortName   string `json:"short_name,omitempty"`
 	Description string `json:"description,omitempty"`
 	CreatedAt   string `json:"created_at,omitempty"`
 	UpdatedAt   string `json:"updated_at,omitempty"`