+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+#### master
+
+- NEW: Added support for Collaborators methods (GH-48)
+- NEW: Added region support to `ZoneRecord` (GH-47)
+- NEW: Added support for Domains Pushes methods (GH-42)
+
+- CHANGED: Renamed `DomainTransferRequest.AuthInfo` to `AuthCode` (GH-46)
+- CHANGED: Updated registration, transfer, renewal response payload (dnsimple/dnsimple-developer#111, dnsimple/dnsimple-go#52).
+
 
 #### Release 0.13.0